Move `get_artifact_url_by_build` into Helpers module
